tests/wpt/tests/domxpath/fn-substring.html

<!DOCTYPE html>
<link rel="help" href="https://www.w3.org/TR/1999/REC-xpath-19991116/#function-substring">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="helpers.js"></script>
<body>
<div id="context"></div>
<script>
test(() => {
  const context = document.querySelector('#context');
  context.innerHTML = '<span>^^^bar$$$</span><span></span><br><br><br><br>';
  assert_equals(evaluateString('substring((./span)[1], count(./br))', context), 'bar$$$');
}, 'substring() arguments depending on the context node');
</script>
</body>
