Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/domxpath/fn-substring-expected.txt

^^^bar$$$




FAIL substring() arguments depending on the context node assert_equals: expected "bar$$$" but got "^^^bar$$$"
